add csrf protection

+++ src/Controller/DnsController.php
@@ -5,6 +5,7 @@ namespace Ws\Controller;
 use WS\Dns\DnsServiceInterface;
 use WS\Dns\RecordFactory\RecordFactory;
 use WS\Http\Request;
+use WS\Session\CSRF;
 use WS\Session\FlashMessage;
 use WS\Template\Template;
 
@@ -23,26 +24,31 @@ class DnsController
 
     public function createAction(Request $request, DnsServiceInterface $dnsService)
     {
+        $csrfGuard = new CSRF();
         $flashMessage = new FlashMessage();
         if ($request->isPost()) {
-            $type = $request->post('type', 'a');
-            $name = $request->post('name');
-            $content = $request->post('content');
-            $ttl = $request->post('ttl', 600);
-            $priority = $request->post('prio');
-            $weight = $request->post('weight');
-            $port = $request->post('port');
+            if ($csrfGuard->validate($request->post('token'))) {
+                $type = $request->post('type', 'a');
+                $name = $request->post('name');
+                $content = $request->post('content');
+                $ttl = $request->post('ttl', 600);
+                $priority = $request->post('prio');
+                $weight = $request->post('weight');
+                $port = $request->post('port');
 
-            $record = RecordFactory::createFromArray(['type' => $type, 'name' => $name, 'content' => $content, 'ttl' => $ttl, 'prio' => $priority, 'weight' => $weight, 'port' => $port]);
-            $apiResult = $dnsService->createRecord($record);
-            if (true === $apiResult) {
-                $flashMessage->addSuccess('Record was created');
-                $this->redirect('/list');
+                $record = RecordFactory::createFromArray(['type' => $type, 'name' => $name, 'content' => $content, 'ttl' => $ttl, 'prio' => $priority, 'weight' => $weight, 'port' => $port]);
+                $apiResult = $dnsService->createRecord($record);
+                if (true === $apiResult) {
+                    $flashMessage->addSuccess('Record was created');
+                    $this->redirect('/list');
+                }
+
+                // fail:
+                $flashMessage->addError('Error during api call - probably incorrect data.');
+                $flashMessage->addError('Error: ' . $apiResult);
+            } else {
+                $flashMessage->addError('Problem with CSRF token');
             }
-
-            // fail:
-            $flashMessage->addError('Error during api call - probably incorrect data.');
-            $flashMessage->addError('Error: '.$apiResult);
         }
 
 
@@ -50,24 +56,31 @@ class DnsController
         $template->menuLabel = 'Back to listing';
         $template->menuLink = '/list';
         $template->flashMessage = $flashMessage;
+        $template->token = $csrfGuard->generate();
 
         echo $template->render();
     }
 
     public function deleteAction(Request $request, DnsServiceInterface $dnsService)
     {
+        $csrfGuard = new CSRF();
         $deleteId = $request->query('id');
 
         if ($request->isPost()) {
             $flashMessage = new FlashMessage();
-            $flashMessage->addSuccess('Record was removed');
-            $dnsService->deleteRecord($deleteId);
-            $this->redirect('/list');
+            if ($csrfGuard->validate($request->post('token'))) {
+                $flashMessage->addSuccess('Record was removed');
+                $dnsService->deleteRecord($deleteId);
+                $this->redirect('/list');
+            } else {
+                $flashMessage->addError('Problem with CSRF token');
+            }
         }
 
         $template = new Template(__DIR__.'/../Resources/templates/delete_confirm.php');
         $template->menuLabel = 'Back to listing';
         $template->menuLink = '/list';
+        $template->token = $csrfGuard->generate();
         $template->id = $deleteId;
 
         echo $template->render();
